Fix buffer overrun in CREATE SCHEMA
Attribute "nspname" of table "sys.babelfish_namespace_ext" is of
type "name", but add_ns_ext_info() did not use a variable of type
NameData for the value.
Now "heap_form_tuple" assumes that new_record[0] is a "name" and
will happily copy NAMEDATALEN bytes, reading past the end of the
string.

Signed-off-by: Laurenz Albe <laurenz.albe@cybertec.at>