Fix buffer overrun in CREATE DATABASE
Attribute "owner" of table "sys.babelfish_sysdatabases" is of type "name",
but "do_create_bbf_db()" did not use a variable of type "NameData" for the value.
Now "heap_form_tuple()" assumes that "new_record[3]" is a "name" and will
happily copy NAMEDATALEN bytes, reading past the end of the string.

Signed-off-by: Laurenz Albe <laurenz.albe@cybertec.at>