Allow multiple references to same column in OUTPUT target list
When generating an output list for the OUTPUT clause, the same column can appear multiple times in the list. 
This happens as different functions/expressions can reference the same columns, leading to a column reference ambiguity error for column aliases ('column reference ... is ambiguous'). As a solution, we do not generate a column alias if the  underlying column is already part of the output list.

Task: BABEL-2522
Author: Avantika Dasgupta <davantik@amazon.com>
Signed-off-by: Harsh Lunagariya <lunharsh@amazon.com>