Create the fixeddecimal_ops operator family and bump up babelfishpg_common version to 1.1.0
This commit supports cross-operator operation (comp/hash operator for index) on fixeddecimal along with INT2/INT4/INT8 properly. 
Before this fix, we relied on operator class but this may cause an error "missing operator in opfamily" because the optimizer might rewrite a qual so optimizer may try to find operator in wrong opfamily. There is an example case that, original join conditon is operator=(fixeddecimal,int), which is included in opfamily fixeddecimal_ops. But it was rewritten to other join condition due to transitive rule so it is rewritten to operator=(int,int) which is not included in given opfamily fixeddecimal_ops.

In this change, we define operator family covering integer/fixeddecimal together as PG's official document guides: https://www.postgresql.org/docs/13/xindex.html#XINDEX-OPFAMILY

Task: BABEL-2089
Author: Sangil Song <sonsangi@amazon.com>
Signed-off-by: Sharu Goel <goelshar@amazon.com>