Make text, next, and image data types invalid for local variables
Previously, we were not raising an error on the declaration of text, ntext and image local variables.

This commit adds a type-check for text, ntext, and image at makeDeclareStmt function, and throws an error as they are not allowed datatypes for a declare statement.  It adds a supportive utility function for checking text, ntext and image datatype OIDs. It also moves is_tsql_xxx_datatypes functions to datatypes.h and implements them in typecode.c & typecode.h in babelfishpg_common extension.

Task: BABEL-2276
Author: Harsh Lunagariya <lunharsh@amazon.com>
Signed-off-by: Avantika Dasgupta <davantik@amazon.com>