Support sp_updatestats system procedure
This commit adds the sys.sp_updatestats procedure, a created system procedure that executes 'ANALYZE' (collects statistics about contents of tables). The procedure performs similarly to sp_updatestats. 

It will not print any messages as output; all progress messages will be in logs as warnings or infos and the resample option will be ignored.

Task: BABEL-1309
Author: Harsh Lunagariya <lunharsh@amazon.com>
Signed-off-by: Sharu Goel <goelshar@amazon.com>