Restricting non-babelfish user from Login through TDS endpoint except superusers
This commit adds restrictions for a user to login through the TDS endpoint.  
The user must be one of the following:
1. The PostgreSQL user that was used to initialize Babelfish using sys.babelfish_initialize() 
procedure call. (master user of babelfish_db).
2. A PostgreSQL superuser.
3. A user created using CREATE LOGIN command through TDS endpoint.

If none of the above conditions is satisfied, then an appropriate error is thrown to let the 
user know that the connecting user is not a Babelfish user.

Also, removed the assert for Babelfish connection object in the JDBC test framework. 
Earlier we were asserting that the connection object should not be NULL. 
If the connection object was NULL, it would cause the assertion to fail and the test file 
won't be executed fully causing the database to be in an inconsistent state. Now if the 
connection object is NULL, only the associated login will fail and won't stop the test execution.

Task: BABEL-2648
Author: Rishabh Tanwar <ritanwar@amazon.com>
Signed-off-by: Harsh Lunagariya <lunharsh@amazon.com>