Apply name resolution to MSTVF's result table type
During function creation, the function name is adjusted to the corresponding 
logical database and schema name. However, when creating a MSTVF, a result 
type associated with it is also created. The result type name lacks the same logic 
and thus throws an error. This commit extends the function name adjusting logic 
to the result type name for MSTVF.

Task: BABEL-2647
Author: Zitao Quan <qztao@amazon.com>
Signed-off-by: Avantika Dasgupta <davantik@amazon.com>