Return variant base-types for NOT NULL constrained columns and handle special cases in Bulk Load Request Support for fixed length datatypes
During Bulk Load Request Testing we ran into an issue where any fixed length type datatype column with a special clause like NOT NULL came with a different type with no max-length in the metadata. The type being sent were exactly same as the sql-variant base types.

To fix this this commit includes two separate fixes:
Adds support on the TDS end to return variant base-types instead of the fixed length datatypes for NOT NULL constrained columns and also avoid sending the max-lengths.
Handled the special cases in Bulk Load Request Support for these fixed length datatypes as the variant base-types in-case of NOT NULL constrained columns.

Moreover in TDS protocol, flags for the not null column come as 0x08, whereas for primary key columns, it is 0x10. We send 0x09 which contains the flag nullable. So we have only fixed generic issue by sending 0x08 which doesn't contain the nullable flag. We are not sending the right value for all types of flags yet.

Task: BABEL-216/2689/2733 
Author: Kushaal Shroff <kushaal@amazon.com>, Shalini Lohia <lshalini@amazon.com>
Signed-off-by: Sharu Goel <goelshar@amazon.com>