Block dropping the last member of sysadmin
syadmin is a built-in role with server level management permissions. It is by default granted to the first login ('sa'). The user can further GRANT/REVOKE it for other login users. 

The symptom of this bug is that user could accidentally revoke all members of sysadmin, which makes the server no-longer manageable. 

This commit fixes the issue by maintaining at least one sysadmin member by rejecting the attempt to drop the last member of sysadmin.

Task: BABEL-2569
Author: Di Wu <amzndiwu@amazon.com>
Signed-off-by: Sharu Goel <goelshar@amazon.com>