Add extension creation script for workflows
Currently these steps are hand-coded into ci.yml. This commit pulls them
out so other test workflows can use them. It's not a true refactoring
since ci.yml still has the hard-coded steps. That's how the original
internal commit was done so I've opted not to alter that here. In the
future, we should more fully refactor all of the workflows.

Author: Kushaal Shroff <kushaal@amazon.com>
Signed-off-by: Jim C. Nasby <nasbyj@amazon.com>