Fix conditional RETURN stmt in MSTVF
Previously we handled MSTVF's return by simply adding a RETURN_TABLE statement to the end of the function. This is wrong because there could be early termination such as a RETURN in a conditional block. 

This commit fixes the issue by instead merging the RETURN_TABLE logic into the RETURN statement for MSTVF.

Task: BABEL-2676
Author: Zitao Quan <qztao@amazon.com>
Signed-off-by: Sharu Goel <goelshar@amazon.com>