Fix error when generating sqlvariant datum from nvarchar in .NET program
Previously in dotNet, when we built a varlena sqlvariant without specifying the base type, we read in a cstring and build a sqlvariant upon it. However the cstring does not contain the necessary varlena header, which would later lead to palloc error because the varlena length info is missing. This commit fixes this problem by adding the correct varlena header when building sqlvariant.

Task: BABEL-2635
Author: Xiaohui Fanhe <hexiaohu@amazon.com>
Signed-off-by: Xiaohui Fanhe <hexiaohu@amazon.com>