Output Parameter of type Binary gives TDS protocol error
For binary [ ( n ) ] - fixed-length binary data with a length of n bytes, where n is a value from 1 through 8,000. The storage size is n bytes. 
However, the default length is 1 when n isn't specified in a data definition or variable declaration statement.
The fix includes setting maxLen = 1 when atttypod = -1

Task: BABEL-1671
Author: Satarupa Biswas <satarupb@amazon.com>
Signed-off-by: Sharu Goel <goelshar@amazon.com>