Handling zero data length for [VAR]CHAR datatype of sql_variant
Prior to this commit, the output of server_to_any() API (the API that
does the encoding conversion) was being passed to strlen() to calculate
the actual data length of any string datatypes. The output of the
server_to_any() API can't be assumed to be NULL terminated which caused
an incorrect length of actual data for some cases and abnormal
behaviour on the TDS side.

With this commit, we add an extra check of length before calling the
server_to_any() API and handle the length of data appropriately. 

Task: BABEL-2794
Author: Dipesh Dhameliya <dddhamel@amazon.com>
Signed-off-by: Sharu Goel <goelshar@amazon.com>