TDS EOF log's context seems to be stale
For EOF on TDS Socket we previously didn’t update the context at the
time of logging the relevant information. This created confusion that
the stale context was an actual threat. Fixed it by adding a context
for EOF logs as well.

This commit also add a context for Bulk Load Request in the TDS layer.

Task: BABEL-2796
Author: Kushaal Shroff <kushaal@amazon.com>
Signed-off-by: Sharu Goel <goelshar@amazon.com>