Invalidate Babelfish catalogs properly in syscache
When PostgreSQL inserts/updates/drops tuples in the catalog, before
adding an invalidation message it will double-check if the target
relation is really a catalog. This invalidation message is executed
during command boundary. As a result, for Babelfish catalogs,
PostgreSQL doesn’t add the invalidation message and those invalid entry
will stay in syscache.

With this commit, we have used Syscache instead of table scan.

Task: BABEL-2578
Author: Shalini Lohia <lshalini@amazon.com>
Signed-off-by: Sharu Goel <goelshar@amazon.com>