Support UPDATE/DELETE of table specified by FROM clause alias
T-SQL syntax allows the target relation of UPDATE/DELETE be an alias of
a table in the FROM clause. This commit adds a step during parsing that
checks if the target relation matches with any alias in the FROM
clause. If it does, we update the target relation to have the actual
table name, with the alias, and remove the FROM clause table's alias to
avoid collision.

Task: BABEL-2565
Author: Zitao Quan <qztao@amazon.com>
Signed-off-by: Sharu Goel <goelshar@amazon.com>