Use T-SQL best-match heuristic to select function or procedure
Two issues are observed during investigation in BABEL-2687:
  1. PostgreSQL's select function logic relies on preferred logic. There is no such 
logic in T-SQL. The function should be chosen by its precedence rule only.
  2. PostgreSQL's best-match heuristic always uses a baseType of domainType. 
In Babelfish, we define some datatypes (i.e. tinyint, smallmoney) as a domain, so 
their operator's output datatype is resolved as its baseType, which is wrong in T-SQL.

To resolve these issues, PostgreSQL's best-match heuristic is copied and adjusted 
into the func_select_candidate_hook(), which addresses the above issues.

This commit also adds:

Logic that will prefer to choose an operator with the same datatype argument. For 
example, if given f(int,smallmoney), and we have f(int,money) and 
f(smallmoney,smallmoney) in the catalog, we will choose f(smallmoney, smallmoney), 
which is closer to T-SQL behavior.

Operators with the same domain type as input arguments. These operators will be 
chosen by func_select_candidate_hook() in T-SQL dialect. In PostgreSQL dialect, 
they will not be chosen since PostgreSQL always looks up baseTypes.

pg_catalog.name and pg_catalog.byte are moved to a precedence table with low 
priority. PostgreSQL's built-in functions (with the same arguments) will be chosen 
only when they are explicitly used.

Task: BABEL-2687
Author: Sangil Song <sonsangi@amazon.com>
Signed-off-by: Justin Jossick <jusjosj@amazon.com>