Fix inconsistent JDBC tests
These tests confirm that the views we created are working fine and
ensures that no error occurs when we SELECT from the views. Previously
these two tests would fail from time to time because we specify the
select result to be more than 1, while in some cases there can be only
one result. This commit fixes this issue.

Task: BABEL-2417/BABEL-2596
Author: Xiaohui Fanhe hexiaohu@amazon.com
Signed-off-by: Dipesh Dhameliya dddhamel@amazon.com