Add a separate escape hatch for IGNORE_DUP_KEY=ON
The IGNORE_DUP_KEY property of a unique index or constraint has
semantics that PostgreSQL does not support. This commit adds an escape
hatch to handle this problem.

When an index or constraint is created with IGNORE_DUP_KEY=ON, we
should provide an escape hatch
(babelfishpg_tsql.escape_hatch_ignore_dup_key) with a default setting
of strict. IGNORE_DUP_KEY=OFF can be silently ignored since this does
not represent a difference in behaviour.

Task: BABEL-2488 
Author: Shalini Lohia <lshalini@amazon.com>
Signed-off-by: Sharu Goel <goelshar@amazon.com>