Show only the backends connected to the Babelfish database in sys.sysprocesses
Previously, we were showing all of the processes (both TDS and non-TDS
backends) as part of sys.sysprocesses view.

In this commit, we change the processes that can be seen in the
sys.sysprocesses view. Only the backends connected to the physical
Babelfish database can now be seen from this view. This is because
sys.sysprocesses view is a T-SQL view and is relevant to Babelfish
only. Thus, it makes sense to only show backends connected to the
database on which Babelfish is running (the Babelfish database).

Also, non-TDS backends will have the "dbid" value as 0 (invalid oid)
since dbid refers to the T-SQL database id (the logical database id in
Babelfish).

Task: BABEL-2829
Author: Sharu Goel <goelshar@amazon.com>
Signed-off-by: Harsh Lunagariya <lunharsh@amazon.com>