UPDATE trigger raises error
Previously, an UPDATE trigger raised the error "variable not found in subplan target list" when an UPDATE is performed.

This commit fixes the problem. In UPDATE/DELETE statements, if the target table appears in the FROM-clause again, T-SQL just ignores it. In PostgreSQL, the target table and the FROM-table are regarded as different tables, so it will apply the cartesian product. To remove the cartesian product, we add an additional self-join condition using ctid (target_table.ctid = from_table.ctid).

Task: BABEL-2884
Author: Zhibai Song <szh@amazon.com>
Signed-off-by: Harsh Lunagariya <lunharsh@amazon.com>