Implement sys.endpoints view in Babelfish
Previously, the sys.endpoints view was missing from Babelfish; the view is required by SSMS as part of its connection process. This commit adds the sys.endpoints view to Babelfish.

Note that, unlike SQL Server, this view only contains one row, which describes the T-SQL endpoint. This is because the other endpoint protocols are not currently supported by Babelfish.

Task: BABEL-2802
Author: Aaron Congo <aarcongo@amazon.com>
Signed-off-by: Xiaohui Fanhe <hexiaohu@amazon.com>