A CHECK constraint violation aborts a batch, when it should abort only a statement if the error is raised during DML operations
Some T-SQL error codes can behave differently depending on when an error is raised or what 
operations were being performed at the time. For example, T-SQL error code 547 can behave in 
two different ways:

1. Error code 547 will behave as a statement terminating error if it is raised when DML was being 
executed.
2. The same error code will behave as a transaction aborting error for any other cases.

The scenario described in point no. 1 is happening probably because another T-SQL error token 
(for example 3621) is also being raised after original token. Babelfish can not raise multiple tokens.

In this commit, a new function named override_txn_behaviour is added that overrides behaviour of 
any error code based on the situation. This function will take pltsql_stmt as an input argument.

For the above mentioned scenario, we get the cached plan source using SPI_plan_get_plan_sources 
API, and then we determine what kind of operation was being performed in case of error using 
cached plan source; we then construct a specific flag that indicates what behaviour is supposed to 
be exhibited.

Now, while determining transactional behaviour of any error, we first check if the override_flag 
(override the transactional behaviour) is set. If it is, we use the flag to determine the behaviour. If 
the flag is not set then we follow the standard behaviour.

Task: BABEL-2880/2881 
Author: Dipesh Dhameliya <dddhamel@amazon.com>
Signed-off-by: Xiaohui Fanhe <hexiaohu@amazon.com>