Delimit PostgreSQL reserved keywords used in table reference and table alias
1. Some PostgreSQL reserved keywords are unreserved keywords in T-SQL. 
To use those keywords in a table name (or alias), they should be delimited. 
We already have similar logic for column name and object name but some 
cases for table name are missing. This commit adds the missing logic properly.

2. Applied "delimit" logic to batch level statements such as CREATE-VIEW. 
Internally, incorporated existing code with a common function used in a normal 
SELECT case. This commit adds tsqlCommonMutator for this purpose.

Task: BABEL-2117
Author: Sangil Song <sonsangi@amazon.com>
Signed-off-by: Xiaohui Fanhe <hexiaohu@amazon.com>