MERGE 095: Re-work on sp_columns_100 (#539)
DMS ran into the following bug where sp_columns_100 behaves differently in master and tempdb and in single-db and multi-db. This was because the physical names were leaking through due to stale usage of PostgreSQL's information_schema.

To fix this we not only had to change the reference to information_schema_tsql, but also add more joins to prevent a leak through to other logical schemas and databases. Moreover, most of the columns held inconsistent data which is fixed with this commit as well.

Added detailed test cases to JDBC as well.

Task: BABEL-2926
Author: Kushaal Shroff kushaal@amazon.com
Signed-off-by: Kushaal Shroff kushaal@amazon.com