Modify ROWVERSION implementation to use an ordinary column filled with 8 byte txid (epoch + xmin)
With this commit we will now physically store complete transaction ID (epoch + xmin) information in the ROWVERSION/TIMESTAMP column of a table. Thus we don't have to worry about wraparound (with 8 byte transactionId) and column reference modifications.

This commit adds the sys.get_current_full_xact_id() function as the default value for ROWVERSION/TIMESTAMP in the user's DDL to achieve this. The function sys.get_current_full_xact_id() is a helper function that returns the current full-transactionId.

Additionally an escape hatch is introduced to disallow creating a ROWVERSION/TIMESTAMP column by default which can be overridden by changing the GUC value to ignore.

Task: BABEL-2894/2916
Author: Rishabh Tanwar ritanwar@amazon.com
Signed-off-by: Kushaal Shroff kushaal@amazon.com