Support T-SQL's information_schema.domains
Implemented the information_schema_tsql.domains view. The view returns one row for each alias data type that can be accessed by the current user in the current database. The alias could be a simple user-defined type from a system datatype or it could be a user-defined table type.

This commit also fixes an inherent problem with the sys.is_table_type which returns NULL for table types created in any schema other than dbo. To fix this, the commit removes the reference to pg_'object'_is_visible type of functions and joined it with sys.schema to filter the shared schemas.

Added detailed test cases for information_schema.domains in JDBC.

Task: BABEL-2918
Author: Kushaal Shroff kushaal@amazon.com
Signed-off-by: Satarupa Biswas satarupb@amazon.com