sysname datatype is missing from sys.types
Prior to this commit, the sys.translate_pg_type_to_tsql function didn't return the sysname datatype since there was no mapping for it. 

This commit adds a mapping in the typecodes.c:type_infos array to allow Babelfish to display the sysname type. The fix required update for sysname in functions where we send datatype-related information.

Updated test-cases in JDBC.

Task: BABEL-2919
Author: Kushaal Shroff <kushaal@amazon.com>
Signed-off-by: Xiaohui Fanhe <hexiaohu@amazon.com>