Set user session properties
This commit internally sets the current user to the database user based on the current database and login. It also adjusts the search path such that the user's default schema is used. If a user is not provided with the given login, then it will default to the dbo user or guest user.

Task: BABEL-2856
Author: Tony Lin <tonyvlin@amazon.com>
Signed-off-by: Justin Jossick <jusjosj@amazon.com>