Add max/min() on (var)char datatypes explicitly
Previously, MIN() and MAX() did not produce a result with the same datatype as their argument. In testing, MIN() or MAX() on an NVARCHAR column produced a result with a datatype of double precision.

With recent work, the function look-up logic to use PostgreSQL's preferred datatype was disabled. Instead, the datatype was determined by a T-SQL precedence rule. 

max/min(pg_catalog.text) was used before but it is now resolved to max/min(pg_catalog.double precision) because there is no exactly matching function for max/min(sys.varchar), and double precision's precedence is higher than that of text. To resolve this issue, add max/min() on (var)char explicitly. This should result in the correct aggregate function being chosen.

Task: BABEL-2964
Author: Sangil Song <sonsangi@amazon.com>
Signed-off-by: Justin Jossick <jusjosj@amazon.com>