Initialise BBF_Unicode_General_* collations properly using value of babelfishpg_tsql.default_locale GUC
The general collation, BBF_Unicode_General_* is specially supported in Babelfish; we use it with locale info from the babelfishpg_tsql.default_locale GUC. BBF_Unicode_General_* collations should be initialised in the init_collid_trans_tab() API, but it was not happening due to an issue in the logic when assessing if a given collation is a general collation. This caused BBF_Unicode_General_* collations to always be initialised with lcid = 1033 and code_page = 0.

This commit fixes the the logic of assessing whether a given collation is one of the BBF_Unicode_General_* collations or not. BBF_Unicode_General_* collations will now be initialised correctly using the value of the babelfishpg_tsql.default_locale GUC.

Task: BABEL-2963
Author: Dipesh Dhameliya dddhamel@amazon.com
Signed-off-by: Dipesh Dhameliya dddhamel@amazon.com