Parser fix: correct query rewriting on multi-byte character
Previously, when rewriting a query in the parser, we relied on the token position (CommonToken::startIndex) to indicate the query string fragment which needed to be replaced. The ANTLR parser converted all input to an std::u32string (utf-32 string) internally and ran the lexer/parser on the result. This indicated that token position was based on the character position, not the byte offset. If a multibyte character was involved, an incorrect fragment of query string was rewritten, and relayed a broken query string to the backend parser.

Solution:
To rewrite a query based on token position, we have to convert a query string to std::u32string first so that the offset will indicate a correct position to be replaced. This commit adds the fix to Babelfish.

Task: BABEL-2968
Author: Sangil Song sonsangi@amazon.com
Signed-off-by: Dipesh Dhameliya dddhamel@amazon.com