Handling a database name that includes a minus sign
Babelfish uses the USE command to switch the database context to database requested in a login request when sending the login acknowledgement. If a minus sign is present in the requested database name, the command will fail with an error ("syntax error near '-' at line 1 and character position 10"), and the user won't be able to log into the database.

This was happening because we simply constructed the command without any identifier handling. One point to note here is that, the requested database name must be verified before the execution of USE .

This commit fixes the issue by enclosing the requested database name inside square brackets (delimited identifier) and then executing the command.

This commit removes the babelfishpg_tds.set_db_session_property GUC, which is no longer being used.

Task: BABEL-2993
Author: Dipesh Dhameliya dddhamel@amazon.com
Signed-off-by: Rishabh Tanwar ritanwar@amazon.com