load_error_mapping() should not be called until backend startup is complete (#670)
API load_error_mapping() is being used to load the error mapping hash table which is being used to map the PostgreSQL error to T-SQL's error code.

Previously, load_error_mapping() was being called to load the error mapping hash table even if backend initialization is not compeleted (before the PGPROC structure and other shared memory is allocated to the backend). This will later cause a segmentation fault and crash the server.

With this commit, Babelfish will try to load error mapping hash table only if we are in state where we can create hash tables safely. Otherwise, we will just skip loading the hash table and throw the default error code.

Task: BABEL-3049
Author: Dipesh Dhameliya dddhamel@amazon.com
Signed-off-by: Justin Jossick jusjosj@amazon.com