Truncate precision of DATETIME[2] to 6
PostgreSQL only supports 6 digits of precision (scale)
for TIMESTAMP. The Babelfish types for DATETIME and
DATETIME2 are based on PostgreSQL's TIMESTAMP type. Some
client drivers (like JDBC) transmit the default precision
of 7 for TDS/TSQL. We need to truncate that to avoid
flooding of the logs with WARN level messages.