Collatable constant expressions need to use CLUSTER_COLLATION_OID
SELECT CASE WHEN 'a' = 'A' THEN 'case-insensitive' ELSE 'case-sensitive' END;
returns 'case-sensitive' when it should return 'case-insensitive'.

More generally, constant expressions are not getting the proper default collation,
and so collation inference fails when an expression has only constant expressions.
This was specified in the design document, but was not implemented. In addition,
two other locations in the code were found to be referring to DEFAULT_COLLATION_OID
when they should have been referring to CLUSTER_COLLATION_OID.

The core functional changes applies the CLUSTER_COLLATION_OID instead of the
DEFAULT_COLLATION_OID. The change was made at a low level so that it also applies
to collated parameters.

This change had widespread consequences because constants and parameters that were
previously collated with a deterministic collation (like the "C" collation) were
for the first time needing to cope with the nondeterministic database collation
(which currently is always case-insensitive and accent-sensitive). Regular expressions
and several core string functions don't work for nondeterministic collations, but
fortunately in nearly all cases regular expressions were being used to find delimiters
or separators such as comma that were not case-sensitive, and so the code be corrected
by explicitly collating with the deterministic "C" collation, which always exists.

This change also implements the missing string functions CHARINDEX and PATINDEX,
and corrects the behavior of STRING_SPLIT so that it is case-insensitive.

Task: BABEL-2851
Author: Jim Finnerty <jfinnert@amazon.com>
Signed-off-by: Jungkook Lee <jungkook@amazon.com>