Avoid starting new transaction for insert into table variables
There are two issue which lead to the problem.
A select on a function internally creates a portal to hold the results.
The portal has its own resource owner which is used for all the
snapshots created during function execution. A transaction event inside
the function will reset resource owner to top transaction. Ideally, we
should reset back to portal resource owner at the earliest. This is not
happening right now as we do not want to do reset for cases where
transaction/savepoint was started before portal creation. We need to
distinguish between these two cases.
We start internal transactions for DDLs/DMLs but not for queries. We
were starting transactions for insert into table variables which we
should not. This internal transaction was resetting the resource owner
from portal to top transaction.

This change fixes the second problem. The first problem will be fixed as
a followup later.

Task: BABEL-3108
Signed-off-by: Surendra Vishnoi <vishnosu@amazon.com>