Smart Github workflows for downloading engine repo
Instead of hard-coding where to download the engine repository from,
create a script that checks for alternate places to download the engine
code from. It first checks for a user's fork of the engine repository,
and if found, sees if there's a branch with the same name as the branch
in the engine repository that kicked off the build. If that doesn't
exist, it then checks for a matching branch in the project engine
repository. If that also doesn't work, it uses the default branch in the
project engine repository.

Signed-off-by: Jim C. Nasby <nasbyj@amazon.com>