[BABEL-938] Support for OPENJSON in Babelfish
This commit implements support for two versions of OPENJSON.

OPENJSON_SIMPLE() - implemented using wrapper functions in
					sys_functions.sql
OPENJSON_WITH() - implemented using C function tsql_openjson_with()
					which uses helper functions added in the engine code
					to retrieve the necessary return values

This commit also modifies the ANTLR parser to rewrite the OPENJSON query
as necessary. In the future, this may be refactored to directly be
supported in the TSQL grammar file instead of relying on ANTLR to rewrite the query.

Signed-off-by: Jason Teng <jasonten@amazon.com>