Configure correct sortId for the collations supported in Babelfish
Previously, sortId field for the most of collations was configured incorrectly. Any non-zero value of sortId might cause
some client to choose different client encoding than the information provided in the lcid of the collation metadata. Due
to this, client might not able to parse some chars correctly and would show them incorrectly to the end user even though
data is correct byte wise.

This commit fixes the issue by configuring sortId to zero (default sort order) for the most of the collations. Hence,
client would use the lcid to determine the client encoding and would show the data correctly to end user.

Task: BABEL-3130
Signed-off-by: Dipesh Dhameliya dddhamel@amazon.com