[BABEL-3090] Support multi-byte client encodings other than UTF-16 for VARCHAR(n)
For the non-Unicode datatypes CHAR, VARCHAR, or TEXT, "n" is interpreted as the maximum number of bytes that may be stored
in the encoding of the collation.  For example, this means that if a Japanese collation is used, then n is the maximum number
of bytes in the SJIS encoding.  PostgreSQL always stores strings in the server encoding, and always interprets "n" as the
maximum number of characters, but it is possible for a string of n characters to require more than n bytes in the client encoding.

With this change, byte length of input would be validated at insertion or update time. And appropriate warning or error would be
raised if needed.

Also, moved encoding conversion functionality from babelfishpg_tds to babelfishpg_tsql because we need byte length checking logic
in TSQL dialect also irrespective of whether client is TDS.