Restore the portal-level snapshot for simple expressions, too
We missed the need of this change to cover pltsql's "simple expression" code path.
 If the first thing we execute after a COMMIT/ROLLBACK is one of those,
rather than a full-fledged SPI command, we must explicitly do
EnsurePortalSnapshotExists() to make sure we have an outer snapshot.
Note that it wouldn't be good enough to just push a snapshot for the duration
of the expression execution: what comes back might be toasted, so we'd better have a snapshot protecting it.

Task: BABEL-3196
Signed-off-by: Shalini Lohia <lshalini@amazon.com>
Co-authored-by: Shalini Lohia <lshalini@amazon.com>