sys.dm_exec_sessions view shows incorrect transaction isolation level boot value
Intially, we were storing the boot value of the
"default_transaction_isolation" GUC in shared memory by first fetching
the value using GetConfigOption() and later converting it into suitable
integer using atoi(). In this particular case, GetConfigOption() was
returning a string that cannot directly be converted as an integer by
atoi. As a result atoi was always returning 0 for the boot GUC value.

This commit uses the PG transaction isolation level mapping of name to
value to convert the string to integer.

Signed-off-by: Sharu Goel <goelshar@amazon.com>