sys.dm_exec_sessions view shows incorrect transaction isolation level boot value
Initially, we were storing the boot value of the
"default_transaction_isolation" GUC in shared memory by first fetching
the value using GetConfigOption() and later converting it into a suitable
integer using atoi(). In this particular case, GetConfigOption() was
returning a string that cannot directly be converted to an integer by
atoi(). As a result atoi() was always returning 0 for the GUC's boot value.

This commit uses the DefaultXactIsoLevel to set the boot value in shared
memory.

Task: BABEL-3209
Signed-off-by: Sharu Goel <goelshar@amazon.com>