Modify sys.indexes view
Because part of the work of sys.indexes was changing column data types,
extensive work was done adding sys.indexes to the
babelfishpg_tsql--1.2.0--1.3.0.sql upgrade script. However, PostgreSQL
does not allow column data types to be changed in views once they
exist, so any addition of updated views to the upgrade script where
data types are changed returned an error. The steps I took to mitigate
this was:
 - in the upgrade script rename sys.indexes to sys.indexes_deprecated.
 - add the updated definition of sys.indexes.
 - add the definition of any views/functions that use sys.indexes as a
   dependency to rebuild them to use the updated sys.indexes definition.
 - if sys.indexes_deprecated has no other dependencies (by, say, a
   customer script) drop the view. Else, keep it for now.

To aid in the above process I added 2 new helper functions:
 - sys.has_dependencies: Returns True if an object is dependant upon
   by other objects. False otherwise.
 - sys.drop_view: If a view is not depended upon by other objects, drop
   the view. Else, leave it.

Task: BABELFISH-343/BABEL-2712
Signed-off-by: Avneet Toor <avneett@bitquilltech.com>