Reflect T-SQL isolation level in sys.dm_exec_sessions view
We reflect the PG default_transaction_isolation GUC in the
transaction_isolation_level column in sys.dm_exec_sessions. However,
this is not correct and we should be reflecting the T-SQL transaction
isolation level in the view.

This commit adds changes to reflect the T-SQL transaction isolation
level in the view even if internally the PG transaction isolation level
might be different.

Task: BABEL-3214
Signed-off-by: Sharu Goel <goelshar@amazon.com>