Outer join terminates connection with error "Internal error at ReadAndHandleColumnData"
Even though a base column is NOT NULL, it can be NULL when projected
from outer part of outer-join. Current logic to send the TDS column
metadata relies on nullability information in catalog because of which
it can happen that TDS extension sends a NULL value for a column it
thinks is NOT nullable. This will result in the extension sending
grabage values as part of response causing clients to terminate
connection.

This commit removes the logic of fetching nullability information from
catalog and instead just treats all columns as NOT nullable. Going
forward, we will need some other method to retrieve teh nullability
information correctly.

Task: BABEL-3141
Signed-off-by: Sharu Goel <goelshar@amazon.com>