Fix duplicate indexes issue in sys.indexes (#202)
In sys.indexes it was noticed that if extra constraints were added to a table with an index, then duplicate indexes are shown when sys.indexes is queried for that table. The issue was caused by LEFT JOIN pg_constraint constr ON constr.conindid = c.oid line in the view definition. The view was modified so that duplicate entries for indexes are no longer returned by removing the join and adding logic to the is_unique_constraint column to simply check if at least one unique constraint exists on that table, and not retrieve all constraints like it did before. 

Also corrected the behaviour of the is_unique_constraint for the view. If the constraint is of type UNIQUE return 1, else 0.

Task: BABELFISH-354
Signed-off-by: Avneet Toor avneett@bitquilltech.com