Correct sys.all_views column types and is_ms_shipped column behaviour #201
The column types for view sys.all_views were changed to the correct types. With this, it was found that the is_ms_shipped column wasn't being set correctly and was hardcoded to be 0. 

To have the is_ms_shipped column set correctly, the sys.all_objects view was modified to now assign the correct value to the is_ms_shipped column for each object: 0 for user objects, and 1 for system objects. The sys.all_objects view will act as the source of truth for all other views that need to set a is_ms_shipped column.

Another view was added, sys.shipped_objects_not_in_sys, which keeps track of system objects that are not under the sys schema. This is used by sys.all_objects to correctly assign a value to the is_ms_shipped column for system objects not under the sys schema.

Task: BABELFISH-350
Signed-off-by: Avneet Toor avneett@bitquilltech.com